+++ integrations/amazon_bedrock/src/haystack_integrations/components/generators/amazon_bedrock/__init__.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 from .chat.chat_generator import AmazonBedrockChatGenerator
-from .chat.utils import print_streaming_chunk
 from .generator import AmazonBedrockGenerator
+from .utils import print_streaming_chunk
 
 __all__ = ["AmazonBedrockChatGenerator", "AmazonBedrockGenerator", "print_streaming_chunk"]
